Move UserStatus to its own cpp/h
