+++ source/ConfigFiles.h
@@ -19,219 +19,6 @@ static bool operator==(Array<T> a1, Array<T> a2) {
 
 
 
-/** Class for handling user status */
-class UserSessionStatus {
-public:
-	String					id = "anon";					///< User ID
-	Array<String>			sessionOrder;					///< Array containing session ordering
-	Array<String>			completedSessions;				///< Array containing all completed session ids for this user
-	static Array<String>	defaultSessionOrder;			///< Default session order
-	static bool				randomizeDefaults;				///< Randomize default session order when applying to individual?
-
-	UserSessionStatus() {}
-
-	/** Load user status from Any */
-	UserSessionStatus(const Any& any) {
-		int settingsVersion = 1; // used to allow different version numbers to be loaded differently
-		AnyTableReader reader(any);
-		reader.getIfPresent("settingsVersion", settingsVersion);
-
-		switch (settingsVersion) {
-		case 1:
-			// Require a user ID
-			reader.get("id", id, "All user status fields must include the user ID!");
-			// Setup default session order, then overwrite if specified	
-			sessionOrder = defaultSessionOrder;
-			if (randomizeDefaults) sessionOrder.randomize();
-			reader.getIfPresent("sessions", sessionOrder);			// Override the default session order if one is provided for this user
-			if (sessionOrder.length() == 0) {						// Check for sessions in list
-				throw format("Must provide \"sessions\" array (or default) for User ID:\"%s\" in user status!", id);
-			}
-			// Get the completed sessions array
-			reader.getIfPresent("completedSessions", completedSessions);
-			break;
-		default:
-			debugPrintf("Settings version '%d' not recognized in UserSessionStatus.\n", settingsVersion);
-			break;
-		}
-	}
-
-	/** Serialize to Any */
-	Any toAny(const bool forceAll = true) const {
-		Any a(Any::TABLE);
-		a["id"] = id;									// populate id
-		a["sessions"] = sessionOrder;					// populate session order
-		a["completedSessions"] = completedSessions; 	// Include updated subject table
-		return a;
-	}
-};
-
-/** Class for representing user status tables */
-class UserStatusTable {
-public:
-	bool allowRepeat = false;							///< Flag for whether to (strictly) sequence these experiments (allow duplicates)
-	bool randomizeDefaults = false;						///< Randomize from default session order when applying to user
-	String currentUser;									///< Currently selected user
-	Array<String> defaultSessionOrder = {};				///< Default session ordering (for all unspecified users)
-	Array<UserSessionStatus> userInfo = {};				///< Array of user status
-
-	UserStatusTable() {}
-
-	/** Load from Any */
-	UserStatusTable(const Any& any) {
-		int settingsVersion = 1; // used to allow different version numbers to be loaded differently
-		AnyTableReader reader(any);
-		reader.getIfPresent("settingsVersion", settingsVersion);
-
-		switch (settingsVersion) {
-		case 1:
-			reader.getIfPresent("currentUser", currentUser);
-			reader.getIfPresent("allowRepeat", allowRepeat);
-			reader.getIfPresent("sessions", defaultSessionOrder);
-			UserSessionStatus::defaultSessionOrder = defaultSessionOrder;				// Set the default order here
-			reader.getIfPresent("randomizeSessionOrder", randomizeDefaults);			
-			UserSessionStatus::randomizeDefaults = randomizeDefaults;					// Set whether default session order is randomized
-			reader.get("users", userInfo, "Issue in the (required) \"users\" array from the user status file!");
-			break;
-		default:
-			debugPrintf("Settings version '%d' not recognized in UserStatus.\n", settingsVersion);
-			break;
-		}
-	}
-
-	/** Serialzie to Any */
-	Any toAny(const bool forceAll = false) const {
-		Any a(Any::TABLE);
-		UserStatusTable def;
-		a["settingsVersion"] = 1;						// Create a version 1 file
-		a["currentUser"] = currentUser;
-		if (forceAll || def.allowRepeat != allowRepeat)				a["allowRepeat"] = allowRepeat;
-		if (forceAll || def.randomizeDefaults != randomizeDefaults)	a["randomizeSessionOrder"] = randomizeDefaults;
-		a["sessions"] = defaultSessionOrder;
-		a["users"] = userInfo;							// Include updated subject table
-		return a;
-	}
-
-	/** Get the user status table from file */
-	static UserStatusTable load(const String& filename) {
-		if (!FileSystem::exists(filename)) {						// if file not found, create a default
-			UserStatusTable defaultStatus = UserStatusTable();		// Create empty status
-			UserSessionStatus user;
-			user.sessionOrder = Array<String> ({ "60Hz", "30Hz" });	// Add "default" sessions we add to
-			defaultStatus.userInfo.append(user);					// Add single "default" user
-			defaultStatus.currentUser = user.id;					// Set "default" user as current user
-			defaultStatus.save(filename);							// Save .any file
-			return defaultStatus;
-		}
-		return Any::fromFile(System::findDataFile(filename));
-	}
-
-	inline void save(const String& filename) { toAny().save(filename);  }
-
-	/** Get a given user's status from the table by ID */
-	shared_ptr<UserSessionStatus> getUserStatus(String id) {
-		for (UserSessionStatus user : userInfo) {
-			if (!user.id.compare(id)) return std::make_shared<UserSessionStatus>(user);
-		}
-		return nullptr;
-	}
-
-	/** Get the next session ID for a given user (by ID) */
-	String getNextSession(String userId = "") {
-		// Return the first valid session that has not been completed
-		if (userId.empty()) { userId = currentUser;  }
-		shared_ptr<UserSessionStatus> status = getUserStatus(userId);
-		// Handle sequence mode here (can be repeats)
-		if (allowRepeat) {
-			int j = 0;
-			for (int i = 0; i < status->sessionOrder.size(); i++) {
-				if (status->completedSessions.size() <= i) {						// If there aren't enough entries in completed sessions to support this
-					return status->sessionOrder[i];
-				}
-				// In the future consider cases where completedSessions doesn't exactly match sessionOrder here... (fine for now?)
-			}
-		}
-		// Default mode here (no repeats)
-		else {
-			for (auto sess : status->sessionOrder) {
-				if (!status->completedSessions.contains(sess)) return sess;
-			}
-		}
-		// If all sessions are complete return empty string
-		return "";
-	}
-
-	/** Add a completed session to a given user's completedSessions array */
-	void addCompletedSession(String userId, String sessId) {
-		for (int i = 0; i < userInfo.length(); i++) {
-			if (!userInfo[i].id.compare(userId)) {
-				userInfo[i].completedSessions.append(sessId);
-			}
-		}
-	}
-
-	void validate(Array<String> sessions, Array<String> users) {
-		bool noSessions = true;	// Flag to demark no sessions are present
-		// Build a string list of valid options for session IDs from the experiment
-		String expSessions = "[";
-		for (String sess : sessions) expSessions += "\"" + sess + "\", ";
-		expSessions = expSessions.substr(0, expSessions.size() - 2);
-		expSessions += "]";
-
-		// Check default sessions for valid ids
-		for(String defSessId : defaultSessionOrder) {
-			noSessions = false;
-			if (!sessions.contains(defSessId)) {
-				throw format("Default session config in user status has session with ID: \"%s\". This session ID does not appear in the experiment config file's \"sessions\" array. Valid options are: %s", defSessId, expSessions);
-			}
-		}
-
-		// Check each user for valid options
-		Array<String> userStatusIds;
-		for (UserSessionStatus userStatus : userInfo) {
-			userStatusIds.append(userStatus.id);
-			// Check all of this user's sessions appear in the session array
-			for (String userSessId : userStatus.sessionOrder) {
-				noSessions = false;
-				if (!sessions.contains(userSessId)) {
-					throw format("User \"%s\" has session with ID: \"%s\" in their User Status \"sessions\" Array. This session ID does not appear in the experiment config file's \"sessions\" array. Valid options are: %s", userStatus.id, userSessId, expSessions);
-				}
-			}
-		}
-
-		// Check current user has a valid config
-		if (currentUser.empty()) {
-			throw "\"currentUser\" field is not specified in the user status file!\nIf you are migrating from an older version of FPSci, please cut the \"currentUser = ...\" line\nfrom the user config file and paste it in the user status file.";
-		}
-		else if (!users.contains(currentUser)) {
-			throw format("Current user \"%s\" does not have a valid entry in the user config file!", currentUser);
-		}
-
-		// Check if no default/per user sessions are present
-		if (noSessions) { 
-			throw "Found no sessions in the user status file!"; 
-		}
-	}
-
-	/** Print the user status table to the log */
-	void printToLog() {
-		for (UserSessionStatus status : userInfo) {
-			String sessOrder = "";
-			for (String sess : status.sessionOrder) {
-				sessOrder += sess + ", ";
-			}
-			sessOrder = sessOrder.substr(0, sessOrder.length() - 2);
-			String completedSess = "";
-			for (String sess : status.completedSessions) {
-				completedSess += sess + ", ";
-			}
-			completedSess = completedSess.substr(0, completedSess.length() - 2);
-
-			logPrintf("Subject ID: %s\nSession Order: [%s]\nCompleted Sessions: [%s]\n", status.id.c_str(), sessOrder.c_str(), completedSess.c_str());
-		}
-	}
-};
-
 /** Weapon configuration class */
 class WeaponConfig {
 public:
