+++ vs/FPSci.lib.vcxproj
@@ -160,6 +160,7 @@
     <ClInclude Include="..\source\GuiElements.h" />
     <ClInclude Include="..\source\Dialogs.h" />
     <ClInclude Include="..\source\UserConfig.h" />
+    <ClInclude Include="..\source\UserStatus.h" />
     <ClInclude Include="..\source\WaypointManager.h" />
     <ClInclude Include="..\source\Weapon.h" />
   </ItemGroup>
@@ -177,6 +178,7 @@
     <ClCompile Include="..\source\TargetEntity.cpp" />
     <ClCompile Include="..\source\GuiElements.cpp" />
     <ClCompile Include="..\source\UserConfig.cpp" />
+    <ClCompile Include="..\source\UserStatus.cpp" />
     <ClCompile Include="..\source\WaypointManager.cpp" />
     <ClCompile Include="..\source\Weapon.cpp" />
   </ItemGroup>
